requirements.txt
streamlit
bcrypt
python-dotenv
pymongo
gridfs
face_recognition
opencv-python-headless
numpy
pillow
